% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epop_ssp2.R
\docType{data}
\name{epop_ssp2}
\alias{epop_ssp2}
\title{Past and projected population sizes for all countries by education under SSP2}
\format{
A data frame with 2,159,766 rows and 7 variables, including:
\describe{
\item{name}{Area name}
\item{country_code}{ISO 3 digit country code}
\item{year}{Year of observation from 1950 to 2100 in five-year steps}
\item{age}{Five-year age groups}
\item{education}{Education group}
\item{sex}{Sex}
\item{epop}{Population size in thousands for each age, sex and education group}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}
}
\usage{
epop_ssp2
}
\description{
A data set containing past and projected population sizes for all countries by education under SSP2
}
\keyword{datasets}
