% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undesa_wpp.R
\name{get_undesa_wpp}
\alias{get_undesa_wpp}
\title{Title}
\usage{
get_undesa_wpp(indicator = "pop", variant = "all")
}
\arguments{
\item{indicator}{Character string based on the \code{indicator} column in the \code{wpp_bulk} data frame, representing the bulk file to be downloaded.}

\item{variant}{Character string based on the \code{variant} column in the \code{wpp_bulk} data frame, representing the bulk file to be downloaded.}
}
\value{

}
\description{
Title
}
