% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loading}
\alias{loading}
\title{Helper function to load a .Rdata inside a function}
\source{
\url{https://stackoverflow.com/a/32065204/513463}
}
\usage{
loading(rdata_file)
}
\arguments{
\item{rdata_file}{A .Rdata file}
}
\value{
An environment
}
\description{
Intended for internal use.
}
\examples{
# not run
# x <- 1:10
# y <- runif(10)
# save(x, y, file = "test.RData")
# data_env <- loading("test.RData")
# ls.str(data_env)
}
\keyword{internal}
