% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcde_location.R
\name{wcde_location}
\alias{wcde_location}
\title{Test if country code or codes are in wic_locations}
\usage{
wcde_location(country_code, version = c("wcde-v3", "wcde-v2", "wcde-v1"))
}
\arguments{
\item{country_code}{\code{vector} of integers representing country codes}
}
\value{
\code{TRUE} if all codes given to \code{country_code} are in wic_locations, \code{FALSE} if one or more are not.
}
\description{
Intended for internal use.
}
\examples{
wcde_location(country_code = c(-11, 44))
wcde_location(country_code = c(100, 44))
wcde_location(country_code = 3)
}
\keyword{internal}
