% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcde_pull_single.R
\name{wcde_pull_single}
\alias{wcde_pull_single}
\title{Pull a single vector for a given measure, scenario and .Rdata file name}
\usage{
wcde_pull_single(measure = NULL, scenario = 2, country_code = NULL)
}
\arguments{
\item{measure}{One character string based on the \code{name} column in the \code{wic_indicators} data frame, representing the variable to be interested.}

\item{scenario}{Vector of length one with a number corresponding the scenarios. See details for more information.}

\item{country_code}{Vector of length one with a country numeric code or label vector.}
}
\value{
A tibble with one column, where the column name matches the value given to \code{measure}
}
\description{
Builds the URL to the Github repo of the Wittgenstein Data Explorer for the given measure, scenario and country code or label vector. Requires a working internet connection. Intended for internal use. Not used anymore as cannot get working with \code{pmap()}
}
\examples{
wcde_pull_single(measure = "tfr", country_code = "900")
}
\keyword{internal}
