% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wic_col4.R
\docType{data}
\name{wic_col4}
\alias{wic_col4}
\title{Colours used in Wittgenstein Human Capital Data Explorer}
\format{
A named vector
}
\usage{
wic_col4
}
\description{
Three sets of colours used for filling education based plots based on the different availabities of detailed education catrogries (four, six or eight groups)
}
\keyword{datasets}
