% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wic_scenarios.R
\docType{data}
\name{wic_scenarios}
\alias{wic_scenarios}
\title{Scenarios used in the Wittgenstein Centre Human Capital Data Explorer}
\format{
A data frame with 9 rows and 6 variables, including:
\describe{
\item{scenario_name}{Full scenario name}
\item{scenario}{Code to match help file of \code{get_wcde} function}
\item{scenario_abb}{Short scenario name}
\item{wcde-v3}{Availability of area in Version 3}
\item{wcde-v2}{Availability of area in Version 2}
\item{wcde-v1}{Availability of area in Version 1}
}
}
\source{
\url{https://dataexplorer.wittgensteincentre.org/}
}
\usage{
wic_scenarios
}
\description{
A data set containing the scenario codes, names short names used in the Wittgenstein Centre Human Capital Data Explorer
}
\keyword{datasets}
