% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past_epop.R
\docType{data}
\name{past_epop}
\alias{past_epop}
\title{Past population sizes for all countries by education}
\format{
A data frame with 574,968 rows and 7 variables, including:
\describe{
\item{name}{Area name}
\item{country_code}{ISO 3 digit country code}
\item{year}{Year of observation from 1950 to 2020 in five-year steps}
\item{age}{Five-year age groups}
\item{education}{Education group}
\item{sex}{Sex}
\item{epop}{Population size in thousands for each age, sex and education group}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/}
}
\usage{
past_epop
}
\description{
A data set containing population sizes for all countries by education between 1950 and 2020 from Version 3.
}
\keyword{datasets}
