% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcde_pull.R
\name{wcde_pull}
\alias{wcde_pull}
\title{Pull multiple vectors for a given indicator, scenarios and .Rdata file names}
\usage{
wcde_pull(indicator = NULL, scenario = 2, country_code = NULL)
}
\arguments{
\item{indicator}{One character string based on the \code{name} column in the \code{wic_indicators} data frame, representing the variable to be interested.}

\item{scenario}{Vector with a numbers corresponding the scenarios. See details in \code{wcde} for more information.}

\item{country_code}{Vector of length one or more of country numeric codes based on ISO 3 digit numeric values.}
}
\value{
A tibble with multiple columns.
}
\description{
Requires a working internet connection. Intended for internal use.
}
\examples{
wcde_pull(indicator = "tfr", country_code = "900")
}
\keyword{internal}
