% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wic_indicators.R
\docType{data}
\name{wic_indicators}
\alias{wic_indicators}
\title{Indicators used in the Wittgenstein Data Explorer}
\format{
A data frame with 31 rows and 8 variables, including:
\describe{
\item{fullname}{Full name of indicator}
\item{name}{Short name of indicator to be used in the \code{measure} argument of the \code{wcde()} function}
\item{age}{Indicator variable for the inclsion of breakdowns by five-year age groups in the measure}
\item{bage}{Indicator variable for the inclsion of breakdowns by broad age groups in the measure}
\item{sage}{Indicator variable for the inclsion of breakdowns by new born as a age group in the measure}
\item{sex}{Indicator variable for the inclsion of breakdowns by sex in the measure}
\item{edu}{Indicator variable for the inclsion of breakdowns by education in the measure}
\item{period}{Indicator variable a period (flow) measure}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}
}
\usage{
wic_indicators
}
\description{
A dataset containing the indicator codes, names and further details used in the Wittgenstein Data Explorer
}
\keyword{datasets}
