% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wic_locations.R
\docType{data}
\name{wic_locations}
\alias{wic_locations}
\title{Locations used in the Wittgenstein Centre Human Capital Data Explorer}
\format{
A data frame with 230 rows and 7 variables, including:
\describe{
\item{name}{Area name}
\item{isono}{ISO 3 digit country code}
\item{continent}{Continent of country}
\item{region}{UN region of country}
\item{dim}{Category or country/region/area}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}
}
\usage{
wic_locations
}
\description{
A dataset containing the location codes, names and further details used in the Wittgenstein Centre Human Capital Data Explorer
}
\keyword{datasets}
