% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epop_ssp2.R, R/wic_scenarios.R
\docType{data}
\name{wic_scenarios}
\alias{wic_scenarios}
\title{Past and projected population sizes for all countries by education under SSP2}
\format{
A data frame with 2,159,766 rows and 7 variables, including:
\describe{
\item{name}{Area name}
\item{country_code}{ISO 3 digit country code}
\item{year}{Year of observation from 1950 to 2100 in five-year steps}
\item{age}{Five-year age groups}
\item{education}{Education group}
\item{sex}{Sex}
\item{epop}{Population size in thousands for each age, sex and education group}
}

A data frame with 5 rows and 3 variables, including:
\describe{
\item{scenario_name}{Full scenario name}
\item{scenario}{Code to match help file of \code{wcde} function}
\item{scenario_abb}{Short scenario name}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}

\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}
}
\usage{
wic_scenarios

wic_scenarios
}
\description{
A data set containing past and projected population sizes for all countries by education under SSP2

A data set containing the scenario codes, names short names used in the Wittgenstein Data Explorer
}
\keyword{datasets}
