% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wic_scenarios.R
\docType{data}
\name{wic_scenarios}
\alias{wic_scenarios}
\title{Scenarios used in the Wittgenstein Data Explorer}
\format{
A data frame with 5 rows and 3 variables, including:
\describe{
\item{scenario_name}{Full scenario name}
\item{scenario}{Code to match help file of \code{wcde} function}
\item{scenario_abb}{Short scenario name}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}
}
\usage{
wic_scenarios
}
\description{
A data set containing the scenario codes, names short names used in the Wittgenstein Data Explorer
}
\keyword{datasets}
